User expenditure get, set, update, delete restrictions

+++ backend/controllers/expenditureController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler')
 const Expenditure = require('../models/expenditureModel')
+const User = require('../models/userModel')
 
 // @desc    Get expenditure
 // @route   GET /api/expenditure
 // @access  Private
 const getExpenditure = asyncHandler(async (req, res) => {
-    const expenditureItems = await Expenditure.find()
+    const expenditureItems = await Expenditure.find({ user: req.user.id })
     res.status(200).json(expenditureItems)
 })
 
@@ -28,7 +29,8 @@ const setExpenditure = asyncHandler(async (req, res) => {
     }
     const expenditure = await Expenditure.create({
         amount: req.body.amount,
-        category: req.body.category
+        category: req.body.category,
+        user: req.user.id,
     })
     res.status(200).json(expenditure)
 })
@@ -41,6 +43,17 @@ const updateExpenditure = asyncHandler(async (req, res) => {
     if (!expenditure) {
         res.status(400)
         throw new Error('Expenditure item not found')
+    }
+    const user = await User.findById(req.user.id)
+    // Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // Make sure the logged in user matches the expenditure user
+    if (expenditure.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
     const updatedExpenditure = await Expenditure.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
@@ -56,6 +69,17 @@ const deleteExpenditure = asyncHandler(async (req, res) => {
     if (!expenditure) {
         res.status(400)
         throw new Error('Expenditure item not found')
+    }
+    const user = await User.findById(req.user.id)
+    // Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // Make sure the logged in user matches the expenditure user
+    if (expenditure.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
     await expenditure.deleteOne()  //remove()
     res.status(200).json({ id: req.params.id })
